+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Description: SO ARM 101 (URDF/MJCF) (thanks to @haixuantao)
 - Add `pull` in the CLI to pull `robot_description` and cache it. (thanks to @haixuantao)
 
 ## [1.18.0] - 2025-06-19